Move dpos-user and use the new types
