+++ src/dpos-user.ts
@@ -13,11 +13,11 @@ import {
   SignedTxMiddleware,
   Contracts,
   EthersSigner
-} from '..'
-import { JSONRPCProtocol } from '../internal/json-rpc-client'
-import { DPOS2, Coin, LoomCoinTransferGateway, AddressMapper } from '../contracts'
-import { IWithdrawalReceipt } from '../contracts/transfer-gateway'
-import { sleep } from '../helpers'
+} from '.'
+import { JSONRPCProtocol } from './internal/json-rpc-client'
+import { DPOS2, Coin, LoomCoinTransferGateway, AddressMapper } from './contracts'
+import { IWithdrawalReceipt } from './contracts/transfer-gateway'
+import { sleep } from './helpers'
 import {
   IValidator,
   ICandidate,
@@ -25,23 +25,26 @@ import {
   LockTimeTier,
   ITotalDelegation,
   ICandidateDelegations
-} from '../contracts/dpos2'
-import { selectProtocol } from '../rpc-client-factory'
-import { overrideReadUrl } from '../client'
+} from './contracts/dpos2'
+import { selectProtocol } from './rpc-client-factory'
+import { overrideReadUrl } from './client'
 
 const log = debug('dpos-user')
 
 const coinMultiplier = new BN(10).pow(new BN(18))
-const ERC20Gateway = require('./contracts/ERC20Gateway.json')
-const ERC20 = require('./contracts/ERC20.json')
+const ERC20GatewayABI = require('./mainnet-contracts/ERC20Gateway.json')
+const ERC20ABI = require('./mainnet-contracts/ERC20.json')
+
+import { ERC20 } from './mainnet-contracts/ERC20'
+import { ERC20Gateway } from './mainnet-contracts/ERC20Gateway'
 
 export class DPOSUser {
   private _wallet: ethers.Signer
   private _client: Client
   private _address: Address
   private _ethAddress: string
-  private _ethereumGateway: ethers.Contract
-  private _ethereumLoom: ethers.Contract
+  private _ethereumGateway: ERC20Gateway
+  private _ethereumLoom: ERC20
   private _dappchainGateway: Contracts.LoomCoinTransferGateway
   private _dappchainLoom: Contracts.Coin
   private _dappchainDPOS: Contracts.DPOS2
@@ -158,8 +161,8 @@ export class DPOSUser {
     this._address = address
     this._ethAddress = ethAddress
     this._client = client
-    this._ethereumGateway = new ethers.Contract(gatewayAddress, ERC20Gateway, wallet)
-    this._ethereumLoom = new ethers.Contract(loomAddress, ERC20, wallet)
+    this._ethereumGateway = new ERC20Gateway(gatewayAddress, ERC20GatewayABI, wallet)
+    this._ethereumLoom = new ERC20(loomAddress, ERC20ABI, wallet)
     this._dappchainGateway = dappchainGateway
     this._dappchainLoom = dappchainLoom
     this._dappchainDPOS = dappchainDPOS
@@ -241,22 +244,22 @@ export class DPOSUser {
    * Deposits funds from mainnet to the gateway
    */
   async depositAsync(amount: BN): Promise<ethers.ContractTransaction> {
-    let currentApproval = await this._ethereumLoom.allowance(
+    let currentApproval = await this._ethereumLoom.functions.allowance(
       await this._wallet.getAddress(),
       this._ethereumGateway.address
     )
-    currentApproval = new BN(currentApproval._hex.split('0x')[1], 16) // ugly way to convert the BN
+    let currentApprovalBN = new BN(currentApproval.toString()) 
 
     log('Current approval:', currentApproval)
-    if (amount.gt(currentApproval)) {
-      let tx: ContractTransaction = await this._ethereumLoom.approve(
+    if (amount.gt(currentApprovalBN)) {
+      let tx: ContractTransaction = await this._ethereumLoom.functions.approve(
         this._ethereumGateway.address,
-        amount.sub(currentApproval).toString()
+        amount.sub(currentApprovalBN).toString()
       )
       await tx.wait()
-      log('Approved an extra', amount.sub(currentApproval))
+      log('Approved an extra', amount.sub(currentApprovalBN))
     }
-    return this._ethereumGateway.depositERC20(amount.toString(), this._ethereumLoom.address)
+    return this._ethereumGateway.functions.depositERC20(amount.toString(), this._ethereumLoom.address)
   }
 
   /**
@@ -402,7 +405,7 @@ export class DPOSUser {
     amount: BN,
     sig: string
   ): Promise<ethers.ContractTransaction> {
-    return this._ethereumGateway.withdrawERC20(amount.toString(), sig, this._ethereumLoom.address)
+    return this._ethereumGateway.functions.withdrawERC20(amount.toString(), sig, this._ethereumLoom.address)
   }
 
   /**
