+++ src/index.ts
@@ -51,7 +51,7 @@ export {
 } from './solidity-helpers'
 export { Entity, IEntityParams } from './plasma-cash/entity'
 export { User as PlasmaUser } from './plasma-cash/user'
-export { DPOSUser } from './dpos/dpos-user'
+export { DPOSUser } from './dpos-user'
 export { SparseMerkleTree, ISparseMerkleTreeLevel } from './plasma-cash/sparse-merkle-tree'
 
 export { PlasmaDB, IDatabaseCoin } from './plasma-cash/db'
